Moved SuggestedCountries Data to dataservices app
