+++ dataservices/models.py
@@ -2,6 +2,7 @@ from django.db import models
 from django.contrib.postgres.fields import JSONField
 
 from core.helpers import TimeStampedModel
+from django.utils.translation import gettext as _
 
 
 class EaseOfDoingBusiness(TimeStampedModel):
@@ -97,3 +98,22 @@ class GDPPerCapita(TimeStampedModel):
 
     def __str__(self):
         return self.country_name
+
+
+class SuggestedCountry(TimeStampedModel):
+    hs_code = models.PositiveIntegerField(_('HS Code'))
+    country = models.ForeignKey(
+        'dataservices.Country',
+        verbose_name=_('Suggested Countries'),
+        on_delete=models.SET_NULL,
+        null=True
+    )
+
+    order = models.PositiveIntegerField(_('Order'), null=True, blank=True)
+
+    def __str__(self):
+        return str(self.hs_code)
+
+    class Meta:
+        verbose_name = 'Suggested Country'
+        verbose_name_plural = 'Suggested Countries'
