+++ dataservices/serializers.py
@@ -47,3 +47,13 @@ class GDPPerCapitalSerializer(serializers.ModelSerializer):
     class Meta:
         model = models.GDPPerCapita
         exclude = ['created', 'id', 'modified']
+
+
+class SuggestedCountrySerializer(serializers.ModelSerializer):
+    country_name = serializers.CharField(source='country__name')
+    country_iso2 = serializers.CharField(source='country__iso2')
+    region = serializers.CharField(source='country__region')
+
+    class Meta:
+        model = models.SuggestedCountry
+        fields = ('hs_code', 'country_name', 'country_iso2', 'region')
