+++ personalisation/views.py
@@ -112,23 +112,3 @@ class RecommendedCountriesView(generics.ListAPIView):
             values('country').\
             annotate(num_countries=Count('country')).\
             order_by('-num_countries')[:10]
-
-
-class SuggestedCountriesView(generics.ListAPIView):
-    serializer_class = serializers.SuggestedCountrySerializer
-    permission_classes = []
-
-    def get_queryset(self):
-        hs_code = self.request.query_params.get('hs_code', '').lower()
-        queryset = models.SuggestedCountry.objects.filter(hs_code=hs_code).\
-            order_by('order').\
-            values('hs_code', 'country__name', 'country__iso2', 'country__region')
-        return queryset
-
-    def get(self, *args, **kwargs):
-        if not self.request.query_params.get('hs_code'):
-            return Response(
-                status=500,
-                data={'error_message': 'hs_code missing in request params'}
-            )
-        return super().get(*args, **kwargs)
